Add files via upload
Demographic studies help to understand the past transformations, present trends, and future
predictions of the social, economic, and political behavior of the population. It plays a vital
role in the decision-making process through statistical analyses and predictions which aids like
a foundation for building a better society.
This report comprises the life expectancy and fertility rates of demographic data from the
International Data Base (IDB) gathered from 227 countries in the years 2001 and 2021. The
purpose of this report is to use descriptive statistical analysis to provide a simplified
interpretation of the variable’s relationship and also to compare and find the correlation between
the variables. For Exploratory Data Analysis (EDA) graphical methods like histogram, bar plot,
scatterplot, box plot, and heat map are used.